Fix warnings on printing pid_t type

+++ helper.c
@@ -260,7 +260,7 @@ int stress_process_dumpable(const bool dumpable)
 	(void)prctl(PR_SET_DUMPABLE,
 		dumpable ? SUID_DUMP_USER : SUID_DUMP_DISABLE);
 #endif
-	snprintf(path, sizeof(path), "/proc/%u/coredump_filter", getpid());
+	snprintf(path, sizeof(path), "/proc/%d/coredump_filter", (int)getpid());
 	if ((fd = open(path, O_WRONLY)) >= 0) {
 		char const *str =
 			dumpable ? "0x33" : "0x00";
